// src/app/theme/pages/default/activities/add-activity/models/athlete.ts
export class Athlete {
    id: Number;
    athlete_public_id: Number;
    first_name: String;
    last_name: String;
    gender: any;
    date_of_birth: any;
    height: Number;
    email: String;
    profile_picture: String;
    sport_id: Number;
    is_team_nl: Boolean;
    level_of_athlete: String;

    address: String;
    is_profile_completed: Boolean;
    is_active: Boolean;
    created_at: Date;
    updated_at: Date;
    unit_id: Number;
    is_sync: Boolean;
    ethnicity:any;

}